Separated ArticleDetail & ArticleFetch Activities
The ArticleDetail Activity handles displaying an Article from the main list of articles.

The ArticleFetch Activity handles the calling of the API via a passed in URL (Intent.EXTRA_TEXT). Once the API call is successful, the Article is automatically stored in the DB. If the user does not save the article it is removed. Else the user saves the article and changes the saved state to true.

Navigating Up ancestrally from the AddTag Activity does not crash the app since the Article KeyID is now being stored in SharedPreferences. Upon returning to either ArticleDetail or ArticleFetch, the Article KeyID is retrieved and the DB ref is pointing to the specified Article.

Added finish() method once an Article is removed (unSaved) from the DB.

Cleaned up code and refactored