Completed Edit Tag Flow
User can now click on Tag to edit it once it is Focused. Edit Icon updates to a Check Icon to let the user know the Tag is being edited.

Once the user clicks on the Check mark the Focus is cleared and the Tag is updated in the DB.

If user click on the Edit Tag Icon, the Focus is moved to the end of the selected Tag in the list and the user can then edit the Tag.

User can also click the Delete button to remove the Tag globally. Will add a check stop notification to make sure user means to delete the Tag permantently.