fix: Fixed Widget From Not Displaying Headlines
The widget was not properly displaying the Article object.

I had to make sure to use getValue(Article.class) when fetching the list of Articles at the specified DataBaseReference

I also limited the body text to 5 lines. This way a small preview of the main Article text is showing in the widget.