refactor: Added KeyId back to Tag
I revereted back to having to use a push key for creating a Tag. It bacame easier to update the name of a tag without having to update the Tag Key as well.

User can now add duplicate Tags. This is not intended behavior. Will need to squash this bug.