feat: Article Features; Share, Webview, Confirm Delete
Share button - User can now share the URL of the article when they click on the share button. New Intent fires off and the system will let the user choose how to resolve the Intent.

Webview - User can now switch to a Webview of the Article. This gives the user more control on how to view the Article.

Confirm Article Deletion - User must now confirm via an Alert Dialog if they, in fact, do want to delete the specified Article. This will help with accidental deletion.