delete: Removed Tag Adapter Class and Small Tag RecyclerView
I wanted to simplify the main Tag and Article classes by removing the Tags Adapter class due to it not being utilized.

Implemented a null check to see if the user is logged in prior to loading the main Tags and Articles list.

FirebaseUtils can now check to see if a Firebase User is signed in.