feat: Definition Dialog Fragment
With the Oxford API hooked up, definitions are now able to be populated in a dialog fragment once the user searches for a word.

refactor: renamed xml and classes for better clarity

network: made sure to implement isNetworkAvailable method to make sure the user has internet prior to fetching Articles or signing in.