feat: Enabled Database Persistence
User can now view content even if internet connection is lost.

ui: new logo for app

chore: LoadingDialogFragment now is modular. Created showLoadingFragment() so other parts of the app can display a loading image. This lets the user know something is happening and to wait.

ux: when the user's article is empty, a View shows a tip to help them save articles.

database: denormalized the Firebase Database. New rules were put in to flatten the database.

New Rules:
 "rules": {
    "users": {
      "$uid": {
        ".read": "$uid === auth.uid",
        ".write": "$uid === auth.uid"
      }
    },
    "articles": {
      "$uid": {
        ".read": "$uid === auth.uid",
        ".write": "$uid === auth.uid"
      }
    },
    "tags": {
      "$uid": {
        ".read": "$uid === auth.uid",
        ".write": "$uid === auth.uid"
      }
    }
  }

feat: Firebase User display name and email are now captured to create a User Profile on the Profile tab. todo: setup the UI for the Profile Tab